mcp-backend/functions/requirements.txt
firebase-admin==6.2.0
firebase-functions==0.4.3
google-auth==2.22.0
mailersend
requests==2.32.3
paypal-server-sdk==0.6.1
reportlab==4.0.4  # Latest stable version
pillow==10.3.0  # Required for image processing in PDFs
opencv-python
weasyprint
python-dotenv
pydantic[email]>=2.0,<3.0
